simplify workflow event and reflex adapter

+++ adapters/reflexstreamer/connect_test.go
@@ -13,7 +13,7 @@ var tables = []string{
 	  id bigint not null auto_increment,
 	  foreign_id varchar(255) not null,
 	  timestamp datetime not null,
-	  type bigint not null default 0,
+	  type int not null default 0,
 	  metadata blob,
 	  
   	  primary key (id)
