+++ adapters/sqlstore/util_test.go
@@ -15,7 +15,7 @@ var migrations = []string{
 		workflow_name           varchar(255) not null,
 		foreign_id             varchar(255) not null,
 		run_id                 varchar(255) not null,
-		status                 bigint not null,
+		status                 int not null,
 		object                 blob not null,
 		is_start               bool not null,
 		is_end                 bool not null,
